[Refector] Post,Comment 클래스 파일 및 코드 이동
