+++ src/gui/component/post/PostRead.java
@@ -1,6 +1,8 @@
-package gui;
+package gui.component.post;
 
+import gui.CustomStyle;
 import gui.component.buttons.*;
+import gui.component.comment.CommentItem;
 import gui.component.global.Header;
 
 import javax.swing.*;
@@ -13,7 +15,7 @@ import java.awt.event.MouseListener;
 import java.util.ArrayList;
 
 public class PostRead extends JFrame {
-
+    ArrayList<CommentItem> commentItems;
     PostRead(){
         setLayout(new BorderLayout());
         UIManager.put("TextArea.inactiveForeground", Color.BLACK);//TextArea enabled(false)인 상태에서도 글씨색 그대로
@@ -295,7 +297,7 @@ public class PostRead extends JFrame {
         commentListWrap.setBackground(CustomStyle.white);
         commentList.add(commentListWrap);
 
-        ArrayList<CommentItem> commentItems = new ArrayList<>();
+        commentItems = new ArrayList<>();
         commentItems.add(new CommentItem("1","1","1"));
         commentItems.add(new CommentItem("2","2","2"));
 
@@ -331,6 +333,11 @@ public class PostRead extends JFrame {
                 commentWriteField.setSize(frameSize.width - (CustomStyle.DISPLAY_MARGIN*2 + 48 + 16+2), 20);
                 //E: comment TextField 크기 변경
 
+                //S: comment 내용 크기 변경
+                commentItems.stream().forEach(
+                        item -> item.setCommentWidth(frameSize.width)
+                );
+                //E: comment 내용 크기 변경
             }
         });
         setBackground(CustomStyle.white);
@@ -340,66 +347,8 @@ public class PostRead extends JFrame {
         //E: JFrame Setting
     }
     
-    //코멘트의 코드가 복잡하고 몇개가 될지 모르므로 따로 클래스로 분리
-    public class CommentItem extends JPanel {
-        CommentItem(String nickname, String date, String txt){
 
-            setBackground(CustomStyle.white);
-            setBorder(BorderFactory.createMatteBorder(0,0,1,0, CustomStyle.borderColor)
-            );
-            setLayout(new BorderLayout());
 
-            JPanel commentItemContent = new JPanel();
-            commentItemContent.setBackground(CustomStyle.white);
-            commentItemContent.setLayout(new BorderLayout());
-            commentItemContent.setBorder(new EmptyBorder(16,16,16,16));
-            add(commentItemContent);
-
-            JPanel commentNicknameWrap = new JPanel();
-            commentNicknameWrap.setBackground(CustomStyle.white);
-            commentNicknameWrap.setLayout(new FlowLayout(FlowLayout.LEFT));
-            commentItemContent.add(commentNicknameWrap, BorderLayout.NORTH);
-
-            JLabel commentNickname = new JLabel(nickname);
-            commentNickname.setFont(CustomStyle.setCutomFont(14, 'b'));
-
-            commentNicknameWrap.add(commentNickname);
-
-            JPanel commentDateWrap = new JPanel();
-            commentDateWrap.setBackground(CustomStyle.white);
-            commentDateWrap.setBorder(new EmptyBorder(-4, 0, 0, 0));
-            commentDateWrap.setLayout(new FlowLayout(FlowLayout.LEFT));
-            commentItemContent.add(commentDateWrap, BorderLayout.CENTER);
-
-            JLabel commentDate = new JLabel(date);
-            commentDate.setFont(CustomStyle.setCutomFont(14, 'n'));
-            commentDate.setForeground(CustomStyle.black50);
-            commentDateWrap.add(commentDate);
-
-            JPanel commentTxtWrap = new JPanel();
-            commentTxtWrap.setBackground(CustomStyle.white);
-            commentTxtWrap.setLayout(new BorderLayout());
-            commentTxtWrap.setBorder(new EmptyBorder(0, 4, 0, 4));
-            commentItemContent.add(commentTxtWrap, BorderLayout.SOUTH);
-
-            JTextArea commentTxt = new JTextArea();
-            commentTxt.setText(txt);
-            commentTxt.setFont(CustomStyle.setCutomFont(16, 'n'));
-            commentTxt.setLineWrap(true);
-            commentTxt.setWrapStyleWord(true);
-            commentTxt.setEnabled(false);
-            commentTxtWrap.add(commentTxt);
-
-            PostRead.this.addComponentListener(new ComponentAdapter(){
-                public void componentResized(ComponentEvent e){
-                    //S: comment 내용 크기 변경
-                    //width: 창 크기 - (창 마진 + post 자체 마진 + textArea 마진 + 부모 선 마진)
-                    commentTxt.setSize(PostRead.this.getWidth() - (CustomStyle.DISPLAY_MARGIN*2 + 48 + 16+2+32), 20);
-                    //E: comment 내용 크기 변경
-                }
-            });
-        }
-    }
 
     public static void main(String[] args) {
         new PostRead();
